# Optimize `Hash` for repeated removals and insertions